7223 grub "setup" built-in broken by 6593 Reviewed by: Richard Lowe <richlowe@richlowe.net> Reviewed by: Toomas Soome <tsoome@me.com> Reviewed by: Robert Mustacchi <rm@joyent.com> Approved by: Richard Lowe <richlowe@richlowe.net>

+++ usr/src/grub/grub-0.97/stage2/fsys_ufs2.c
@@ -311,7 +311,7 @@ loop:
 }
 
 int
-ufs2_embed (int *start_sector, int needed_sectors)
+ufs2_embed (unsigned long long *start_sector, int needed_sectors)
 {
   /* XXX: I don't know if this is really correct. Someone who is
      familiar with BSD should check for this.  */
